Read-only support of H100i Platinum

+++ include/protocol/asetek.h
@@ -84,7 +84,7 @@ enum Asetek_PumpModes
 
 int
 corsairlink_asetek_device_id(
-    struct corsair_device_info* dev, struct libusb_device_handle* handle, uint8_t* device_id );
+    struct corsair_device_info* dev, struct libusb_device_handle* handle, uint16_t* device_id );
 
 int
 corsairlink_asetek_name(
