+++ protocol/rmi/core.c
@@ -51,7 +51,7 @@ convert_bytes_double( uint16_t v16 )
 
 int
 corsairlink_rmi_device_id(
-    struct corsair_device_info* dev, struct libusb_device_handle* handle, uint8_t* device_id )
+    struct corsair_device_info* dev, struct libusb_device_handle* handle, uint16_t* device_id )
 {
     // memset(device_id, 0x00, 1);
     ( *device_id ) = 0xFF;
